requirements.txt
bottle
httplib2
simplejson
python-instagram
